PatentMME/requirements.txt
einops==0.7.0
numpy==1.26.4
omegaconf==2.3.0
Pillow
pytorch_lightning==2.1.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
wandb==0.16.1
-e ./ocr-vqgan
